requirements.txt
pandas==2.0.3
numpy==1.24.3
requests==2.31.0
yfinance==0.2.28
psycopg2-binary==2.9.7
pyspark==3.5.0
scikit-learn==1.3.0
tensorflow==2.13.0
dash==2.14.0
python-dotenv==1.0.0
pyarrow==14.0.1
pytest-spark==0.6.0
fbprophet==1.1.5
plotly==5.18.0
holidays==0.36
scikit-learn>=1.0.0
xgboost>=1.5.0
matplotlib>=3.4.0
joblib>=1.0.0
python-dotenv>=0.19.0
pylint>=2.8.0
pytest>=6.2.5